bit_vector

+++ bit_vector.h
@@ -17,8 +17,8 @@ class bit_vector
 template<size_t SZ>
 bit_vector<SZ>::bit_vector()
 {
-	int count = (SZ - 1) / sizeof(long) * 8;
-	m_num = new long[count];
+	int count = (SZ - 1) / sizeof(size_t) * 8;
+	m_num = new size_t[count];
 }
 
 template<size_t SZ>
@@ -35,7 +35,7 @@ void bit_vector<SZ>::set(size_t index)
 		std::cerr << "Out of range!";
 		return;
 	}
-	m_num[index / (sizeof(long) * 8)] |= 1 << (index % (sizeof(long) * 8));
+	m_num[index / (sizeof(size_t) * 8)] |= 1 << (index % (sizeof(size_t) * 8));
 }
 
 template<size_t SZ>
@@ -46,7 +46,7 @@ void bit_vector<SZ>::reset(size_t index)
 		std::cerr << "Out of range!";
 		return;
 	}
-	m_num[index / (sizeof(long) * 8)] &= ~(1 << (index % (sizeof(long) * 8)));
+	m_num[index / (sizeof(size_t) * 8)] &= ~(1 << (index % (sizeof(size_t) * 8)));
 }
 
 #endif
